# openapi.yaml
openapi: 3.0.3
info:
  title: tr-engine API
  version: 0.9.0
  contact:
    name: tr-engine
    url: https://github.com/LumenPrima/tr-engine
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Backend service for trunk-recorder data ingestion and querying.
    Provides REST APIs for accessing radio system data, calls, talkgroups,
    units, and recorder state.

    ## Authentication

    All endpoints (except `/health` and `/auth-init`) require a bearer token
    via the `Authorization: Bearer <token>` header or `?token=` query param.

    **Two-tier auth (optional):** When `WRITE_TOKEN` is configured, mutating
    operations (POST, PATCH, PUT, DELETE) require the write token. Read
    operations (GET) accept either the read token (`AUTH_TOKEN`) or write
    token. If `WRITE_TOKEN` is not set, all operations accept `AUTH_TOKEN`
    (backwards compatible).

    The `/auth-init` endpoint returns only the read token — the write token
    is never exposed by any endpoint.

    ## Data Model: Systems and Sites

    The data model separates **logical radio networks** from **recording
    sites**:

    - **System** (`system_id`): a logical radio network. For P25, this
      is the WACN-level network identified by `(sysid, wacn)`. For
      conventional, it is identified by `(instance_id, sys_name)`.
      Talkgroups and units belong to the system level because they are
      shared across all sites on the same P25 network.

...

    Example:
...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      operationId: listUsers
      summary: List all users
      description: Returns all user accounts. Admin only.
      tags: [auth]
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  total:
                    type: integer
        "403":
          description: Admin access required
    post:
      operationId: createUser
      summary: Create user
      description: Creates a new user account. Admin only.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                  minLength: 8
                role:
                  type: string
                  enum: [viewer, editor, admin]
                  default: viewer
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Admin access required
        "409":
          description: Username already exists
  /users/{id}:
    patch:
      operationId: updateUser
      summary: Update user
      description: |
        Partially update a user account (role, password, enabled status).
        Admin only. Cannot demote or disable the last admin user.
        Cannot demote or disable your own account.
      tags: [auth]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [viewer, editor, admin]
                password:
                  type: string
                  minLength: 8
                enabled:
                  type: boolean
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Admin access required, or last admin protection
        "404":
          description: User not found
    delete:
      operationId: deleteUser
      summary: Delete user
      description: |
        Deletes a user and cascades to their sessions and API keys.
        Admin only. Cannot delete yourself or the last admin user.
      tags: [auth]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
        "403":
          description: Admin access required, or last admin / self-delete protection
        "404":
          description: User not found

  # ----------------------------------------------------------
  # Systems
  # ----------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token authentication. Pass token in the Authorization header:
        `Authorization: Bearer <token>`

  # ----------------------------------------------------------
  # Reusable Parameters
  # ----------------------------------------------------------
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [viewer, editor, admin]
        enabled:
          type: boolean
        display_name:
          type: string
        last_login:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          description: Machine-readable error code for programmatic handling
          example: not_found
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - internal_error
            - service_unavailable
            - invalid_body
            - invalid_parameter
            - invalid_time_range
            - query_failed
            - ambiguous_id
            - duplicate
            - request_timeout
        error:
          type: string
          example: Resource not found
        detail:
          type: string
          description: Additional context when available
...